// src/app/models/aso/GetAsoModel.ts
export class GetAsoModel{
    id_aso: string = '';
    dia_hora: string = '';
    cnpj: string = '';
    razaosocial: string = '';
    nomefantasia: string = '';
    cep: string = '';
    logradouro: string = '';
    complemento: string = '';
    numero: string = '';
    bairro: string = '';
    localidade: string = '';
    uf: string = '';
    nome: string = '';
    data_nascimento_funcionario: string = '';
    cpf: string = '';
    rg: string = '';
    funcao: string = '';
    setor: string = '';
    nome_profissionalsaude: string = '';
    cpf1: string = '';
    formacao_saude: string = '';
    conselho: string = '';
    registro: string = '';
    estado: string = '';
    medico_trabalho: string = '';
    crm_medico_trabalho: string = '';
    cpf_medico_trabalho: string = '';
    tipo_aso: string = '';
    exame_clinico: string = '';
    data_exame_clinico: string = '';
    acuidadevisual: string = '';
    data_acuidadevisual: string = '';
    avaliacao_psicossocial: string = '';
    data_avaliacao_psicossocial: string = '';
    eletrocardiograma: string = '';
    data_eletrocardiograma: string = '';
    eletroencefalograma: string = '';
    data_eletroencefalograma: string = '';
    espirometria: string = '';
    data_espirometria: string = '';
    glicemia: string = '';
    data_glicemia: string = '';
    audiometria: string = '';
    data_audiometria: string = '';
    fator_rh: string = '';
    data_fator_rh: string = '';
    grupo_sanguineo: string = '';
    data_grupo_sanguineo: string = '';
    hemograma_completo: string = '';
    data_hemograma_completo: string = '';
    testeromberg: string = '';
    data_testeromberg: string = '';
    raio_x_lombo_sacra: string = '';
    data_raio_x_lombo_sacra: string = '';
    raio_x_joelho_direito: string = '';
    data_raio_x_joelho_direito: string = '';
    raio_x_joelho_esquerdo: string = '';
    data_raio_x_joelho_esquerdo: string = '';
    raio_x_joelho_coxofemorais_direito: string = '';
    data_raio_x_joelho_coxofemorais_direito: string = '';
    raio_x_joelho_coxofemorais_esquerdo: string = '';
    data_raio_x_joelho_coxofemorais_esquerdo: string = '';
    raio_x_ombro: string = '';
    data_raio_x_ombro: string = '';
    acido_s_fenil_mercapturico: string = '';
    data_acido_s_fenilmercapturico: string = '';
    acido_s_acido_hipurico: string = '';
    data_acido_hipurico: string = '';
    acido_trans_muconico: string = '';
    data_acido_trans_muconico: string = '';
    parecer_medico: string = '';
    parecer_medico_altura: string = '';
    parecer_medico_confinado: string = '';
    parecer_medico_atividade_subaquatica: string = '';
    observacoes: string = '';

}